+++ CHANGELOG.md
@@ -14,7 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 - Define input/output ports
 - Update execution report
-- 
+
 ### Changed
 
 - Forward cmem-plugin-template to v7.1.0