+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/)
 
 
+## [Unreleased]
+
+### Changed
+- Match only files from current project
+
 ## [3.0.0] 2025-02-13
 
 ### Fixed