+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/)
 
+## [3.1.0] 2025-07-17
+
+### Changed
+
+- Forward cmem-plugin-template to v7.3.0
+
 
 ## [3.0.0] 2025-02-13